Rename hydrate→deserialize, add_→populate_, merge enrich+redact into render
- Rename hydrate_conversation → deserialize_conversation (factory.py)
- Rename add_speaker_names → populate_speaker_names (render.py)
- Rename add_folder_names → populate_folder_names (render.py)
- Merge enrich.py and redact.py into render.py (single output-prep module)
- Delete enrich.py and redact.py
- Use lazy imports for database modules in populate functions
- Update all 29 consumer files (imports + call sites)
- Update all 4 test files
