+++ backend/routers/conversations.py
@@ -24,8 +24,8 @@ from models.conversation import (
     UpdateActionItemDescriptionRequest,
     UpdateSegmentTextRequest,
 )
-from utils.conversations.factory import hydrate_conversation
-from utils.conversations.redact import redact_conversations_for_list
+from utils.conversations.factory import deserialize_conversation
+from utils.conversations.render import redact_conversations_for_list
 from utils.conversations.render import conversation_to_dict
 from models.conversation_enums import ConversationStatus, ConversationVisibility
 from models.conversation_photo import ConversationPhoto
@@ -74,7 +74,7 @@ def process_in_progress_conversation(
         raise HTTPException(status_code=404, detail="Conversation in progress not found")
     redis_db.remove_in_progress_conversation_id(uid)
 
-    conversation = hydrate_conversation(conversation)
+    conversation = deserialize_conversation(conversation)
 
     # Inject calendar context if provided
     if request and request.calendar_meeting_context:
@@ -110,7 +110,7 @@ def reprocess_conversation(
     :return: The updated conversation after reprocessing.
     """
     conversation = _get_valid_conversation_by_id(uid, conversation_id)
-    conversation = hydrate_conversation(conversation)
+    conversation = deserialize_conversation(conversation)
     if not language_code:
         language_code = conversation.language or 'en'
 
@@ -248,7 +248,7 @@ def set_conversation_events_state(
     conversation_id: str, data: SetConversationEventsStateRequest, uid: str = Depends(auth.get_current_user_uid)
 ):
     conversation = _get_valid_conversation_by_id(uid, conversation_id)
-    conversation = hydrate_conversation(conversation)
+    conversation = deserialize_conversation(conversation)
     events = conversation.structured.events
     for i, event_idx in enumerate(data.events_idx):
         if event_idx >= len(events):
@@ -264,7 +264,7 @@ def set_action_item_status(
     data: SetConversationActionItemsStateRequest, conversation_id: str, uid=Depends(auth.get_current_user_uid)
 ):
     conversation = _get_valid_conversation_by_id(uid, conversation_id)
-    conversation = hydrate_conversation(conversation)
+    conversation = deserialize_conversation(conversation)
     action_items = conversation.structured.action_items
     for i, action_item_idx in enumerate(data.items_idx):
         if action_item_idx >= len(action_items):
@@ -325,7 +325,7 @@ def update_action_item_description(
     conversation_id: str, data: UpdateActionItemDescriptionRequest, uid=Depends(auth.get_current_user_uid)
 ):
     conversation = _get_valid_conversation_by_id(uid, conversation_id)
-    conversation = hydrate_conversation(conversation)
+    conversation = deserialize_conversation(conversation)
     action_items = conversation.structured.action_items
 
     found_item = False
@@ -356,7 +356,7 @@ def update_action_item_description(
 @router.delete("/v1/conversations/{conversation_id}/action-items", response_model=dict, tags=['conversations'])
 def delete_action_item(data: DeleteActionItemRequest, conversation_id: str, uid=Depends(auth.get_current_user_uid)):
     conversation = _get_valid_conversation_by_id(uid, conversation_id)
-    conversation = hydrate_conversation(conversation)
+    conversation = deserialize_conversation(conversation)
     action_items = conversation.structured.action_items
     updated_action_items = [item for item in action_items if not (item.description == data.description)]
     conversations_db.update_conversation_action_items(
@@ -409,7 +409,7 @@ def set_assignee_conversation_segment(
         f'set_assignee_conversation_segment {conversation_id} {segment_idx} {assign_type} {value} {use_for_speech_training} {uid}'
     )
     conversation = _get_valid_conversation_by_id(uid, conversation_id)
-    conversation = hydrate_conversation(conversation)
+    conversation = deserialize_conversation(conversation)
 
     if value == 'null':
         value = None
@@ -478,7 +478,7 @@ def set_assignee_conversation_segment(
         f'set_assignee_conversation_segment {conversation_id} {speaker_id} {assign_type} {value} {use_for_speech_training} {uid}'
     )
     conversation = _get_valid_conversation_by_id(uid, conversation_id)
-    conversation = hydrate_conversation(conversation)
+    conversation = deserialize_conversation(conversation)
 
     if value == 'null':
         value = None
@@ -537,7 +537,7 @@ def assign_segments_bulk(
     uid: str = Depends(auth.get_current_user_uid),
 ):
     conversation = _get_valid_conversation_by_id(uid, conversation_id)
-    conversation = hydrate_conversation(conversation)
+    conversation = deserialize_conversation(conversation)
 
     value = data.value
     if value == 'null':
@@ -614,7 +614,7 @@ def get_shared_conversation_by_id(conversation_id: str):
     visibility = conversation.get('visibility', ConversationVisibility.private)
     if not visibility or visibility == ConversationVisibility.private:
         raise HTTPException(status_code=404, detail="Conversation is private")
-    conversation = hydrate_conversation(conversation)
+    conversation = deserialize_conversation(conversation)
     conversation.geolocation = None
 
     # Fetch people data for speaker names
@@ -679,7 +679,7 @@ def get_conversation_suggested_apps(conversation_id: str, uid: str = Depends(aut
     from models.app import App
 
     conversation_data = _get_valid_conversation_by_id(uid, conversation_id)
-    conversation = hydrate_conversation(conversation_data)
+    conversation = deserialize_conversation(conversation_data)
 
     # Get suggested app models with full data (similar to /v1/apps endpoint)
     suggested_apps = []
@@ -714,7 +714,7 @@ def test_prompt(
     uid: str = Depends(auth.with_rate_limit(auth.get_current_user_uid, "test:prompt")),
 ):
     conversation_data = _get_valid_conversation_by_id(uid, conversation_id)
-    conversation = hydrate_conversation(conversation_data)
+    conversation = deserialize_conversation(conversation_data)
 
     full_transcript = "\n".join([seg.text for seg in conversation.transcript_segments if seg.text])
