+++ backend/tests/unit/test_folder_name_enrichment.py
@@ -98,7 +98,7 @@ _stub_module("utils.scopes")
 sys.modules["utils.scopes"].AVAILABLE_SCOPES = {}
 sys.modules["utils.scopes"].validate_scopes = MagicMock()
 
-# utils.conversations.enrich imports database.folders and database.users
+# utils.conversations.render imports database.folders and database.users
 # which are already stubbed above — no additional stubs needed.
 
 _stub_module("utils.llm")
@@ -120,7 +120,7 @@ sys.modules["dependencies"].get_uid_with_goals_write = MagicMock()
 # ---------------------------------------------------------------------------
 # Now import the actual functions under test
 # ---------------------------------------------------------------------------
-from utils.conversations.enrich import add_folder_names, add_speaker_names
+from utils.conversations.render import populate_folder_names, populate_speaker_names
 
 # ---------------------------------------------------------------------------
 # Tests
@@ -128,7 +128,7 @@ from utils.conversations.enrich import add_folder_names, add_speaker_names
 
 
 class TestAddFolderNamesToConversations:
-    """Tests for add_folder_names in developer API."""
+    """Tests for populate_folder_names in developer API."""
 
     def setup_method(self):
         _mock_get_folders.reset_mock()
@@ -142,7 +142,7 @@ class TestAddFolderNamesToConversations:
             {'id': 'conv1', 'folder_id': 'folder1'},
             {'id': 'conv2', 'folder_id': 'folder2'},
         ]
-        add_folder_names('uid1', conversations)
+        populate_folder_names('uid1', conversations)
 
         assert conversations[0]['folder_name'] == 'Work'
         assert conversations[1]['folder_name'] == 'Personal'
@@ -153,7 +153,7 @@ class TestAddFolderNamesToConversations:
             {'id': 'conv1', 'folder_id': None},
             {'id': 'conv2'},
         ]
-        add_folder_names('uid1', conversations)
+        populate_folder_names('uid1', conversations)
 
         assert conversations[0]['folder_name'] is None
         assert conversations[1]['folder_name'] is None
@@ -166,7 +166,7 @@ class TestAddFolderNamesToConversations:
         conversations = [
             {'id': 'conv1', 'folder_id': 'deleted_folder'},
         ]
-        add_folder_names('uid1', conversations)
+        populate_folder_names('uid1', conversations)
 
         assert conversations[0]['folder_name'] is None
 
@@ -179,7 +179,7 @@ class TestAddFolderNamesToConversations:
             {'id': 'conv2', 'folder_id': None},
             {'id': 'conv3'},
         ]
-        add_folder_names('uid1', conversations)
+        populate_folder_names('uid1', conversations)
 
         assert conversations[0]['folder_name'] == 'Work'
         assert conversations[1]['folder_name'] is None
@@ -187,7 +187,7 @@ class TestAddFolderNamesToConversations:
 
     def test_empty_conversations_list(self):
         conversations = []
-        add_folder_names('uid1', conversations)
+        populate_folder_names('uid1', conversations)
 
         assert conversations == []
         _mock_get_folders.assert_not_called()
@@ -203,13 +203,13 @@ class TestAddFolderNamesToConversations:
             {'id': 'conv2', 'folder_id': 'f2'},
             {'id': 'conv3', 'folder_id': 'f1'},
         ]
-        add_folder_names('uid1', conversations)
+        populate_folder_names('uid1', conversations)
 
         assert _mock_get_folders.call_count == 1
 
 
 class TestAddFolderNamesWebhookPayload:
-    """Tests for add_folder_names with single-item webhook payloads."""
+    """Tests for populate_folder_names with single-item webhook payloads."""
 
     def setup_method(self):
         _mock_get_folders.reset_mock()
@@ -218,7 +218,7 @@ class TestAddFolderNamesWebhookPayload:
         _mock_get_folders.return_value = [{'id': 'folder1', 'name': 'Work'}]
         payload = {'folder_id': 'folder1'}
 
-        add_folder_names('uid1', [payload])
+        populate_folder_names('uid1', [payload])
 
         assert payload['folder_name'] == 'Work'
         _mock_get_folders.assert_called_once_with('uid1')
@@ -226,7 +226,7 @@ class TestAddFolderNamesWebhookPayload:
     def test_payload_without_folder_id_gets_none(self):
         payload = {'folder_id': None}
 
-        add_folder_names('uid1', [payload])
+        populate_folder_names('uid1', [payload])
 
         assert payload['folder_name'] is None
         _mock_get_folders.assert_not_called()
@@ -234,7 +234,7 @@ class TestAddFolderNamesWebhookPayload:
     def test_payload_missing_folder_id_key_gets_none(self):
         payload = {}
 
-        add_folder_names('uid1', [payload])
+        populate_folder_names('uid1', [payload])
 
         assert payload['folder_name'] is None
         _mock_get_folders.assert_not_called()
@@ -243,7 +243,7 @@ class TestAddFolderNamesWebhookPayload:
         _mock_get_folders.return_value = []
         payload = {'folder_id': 'deleted_folder'}
 
-        add_folder_names('uid1', [payload])
+        populate_folder_names('uid1', [payload])
 
         assert payload['folder_name'] is None
 
@@ -256,7 +256,7 @@ _mock_get_people_by_ids = users_mod.get_people_by_ids
 
 
 class TestAddSpeakerNames:
-    """Tests for add_speaker_names enrichment."""
+    """Tests for populate_speaker_names enrichment."""
 
     def setup_method(self):
         _mock_get_user_profile.reset_mock()
@@ -266,7 +266,7 @@ class TestAddSpeakerNames:
 
     def test_user_segments_get_user_name(self):
         conversations = [{'transcript_segments': [{'text': 'hi', 'is_user': True, 'speaker_id': 0}]}]
-        add_speaker_names('uid1', conversations)
+        populate_speaker_names('uid1', conversations)
         assert conversations[0]['transcript_segments'][0]['speaker_name'] == 'TestUser'
 
     def test_known_person_id_gets_person_name(self):
@@ -274,7 +274,7 @@ class TestAddSpeakerNames:
         conversations = [
             {'transcript_segments': [{'text': 'hi', 'is_user': False, 'person_id': 'p1', 'speaker_id': 1}]}
         ]
-        add_speaker_names('uid1', conversations)
+        populate_speaker_names('uid1', conversations)
         assert conversations[0]['transcript_segments'][0]['speaker_name'] == 'Alice'
 
     def test_unknown_person_id_falls_back_to_speaker_id(self):
@@ -282,33 +282,33 @@ class TestAddSpeakerNames:
         conversations = [
             {'transcript_segments': [{'text': 'hi', 'is_user': False, 'person_id': 'unknown_p', 'speaker_id': 3}]}
         ]
-        add_speaker_names('uid1', conversations)
+        populate_speaker_names('uid1', conversations)
         assert conversations[0]['transcript_segments'][0]['speaker_name'] == 'Speaker 3'
 
     def test_no_person_id_no_is_user_falls_back_to_speaker_id(self):
         conversations = [{'transcript_segments': [{'text': 'hi', 'speaker_id': 2}]}]
-        add_speaker_names('uid1', conversations)
+        populate_speaker_names('uid1', conversations)
         assert conversations[0]['transcript_segments'][0]['speaker_name'] == 'Speaker 2'
 
     def test_missing_speaker_id_defaults_to_zero(self):
         conversations = [{'transcript_segments': [{'text': 'hi'}]}]
-        add_speaker_names('uid1', conversations)
+        populate_speaker_names('uid1', conversations)
         assert conversations[0]['transcript_segments'][0]['speaker_name'] == 'Speaker 0'
 
     def test_user_profile_missing_name_falls_back_to_user(self):
         _mock_get_user_profile.return_value = {"name": None}
         conversations = [{'transcript_segments': [{'text': 'hi', 'is_user': True, 'speaker_id': 0}]}]
-        add_speaker_names('uid1', conversations)
+        populate_speaker_names('uid1', conversations)
         assert conversations[0]['transcript_segments'][0]['speaker_name'] == 'User'
 
     def test_empty_conversations_list(self):
-        add_speaker_names('uid1', [])
+        populate_speaker_names('uid1', [])
         _mock_get_user_profile.assert_called_once_with('uid1')
         _mock_get_people_by_ids.assert_not_called()
 
     def test_no_transcript_segments_key(self):
         conversations = [{'id': 'conv1'}]
-        add_speaker_names('uid1', conversations)
+        populate_speaker_names('uid1', conversations)
         # Should not crash — no segments to enrich
 
     def test_batch_loads_people_once(self):
@@ -329,7 +329,7 @@ class TestAddSpeakerNames:
                 ]
             },
         ]
-        add_speaker_names('uid1', conversations)
+        populate_speaker_names('uid1', conversations)
         assert _mock_get_people_by_ids.call_count == 1
         assert conversations[0]['transcript_segments'][0]['speaker_name'] == 'Alice'
         assert conversations[1]['transcript_segments'][0]['speaker_name'] == 'Bob'
