+++ backend/utils/conversations/process_conversation.py
@@ -34,7 +34,7 @@ from models.conversation import (
     ExternalIntegrationCreateConversation,
 )
 from models.conversation_enums import ConversationSource, ConversationStatus, ExternalIntegrationConversationSource
-from utils.conversations.factory import hydrate_conversation
+from utils.conversations.factory import deserialize_conversation
 from models.other import Person
 from models.structured import Structured
 from utils.notifications import send_important_conversation_message
@@ -907,7 +907,7 @@ def process_user_expression_measurement_callback(provider: str, request_id: str,
         logger.warning(f"Conversation is not found. Uid: {uid}. Conversation: {task.memory_id}")
         return
 
-    conversation = hydrate_conversation(conversation_data)
+    conversation = deserialize_conversation(conversation_data)
 
     # Get prediction
     predictions = callback.predictions
