+++ backend/utils/conversations/render.py
@@ -9,6 +9,124 @@ if TYPE_CHECKING:
     from models.conversation import Conversation
 
 
+# ---------------------------------------------------------------------------
+# Populate: speaker names, folder names
+# ---------------------------------------------------------------------------
+
+
+def populate_speaker_names(uid: str, conversations: List[Dict]) -> None:
+    """Add speaker_name to transcript segments based on person_id mappings.
+
+    Mutates conversation dicts in-place. Works with both single conversations
+    (pass as [conv]) and lists.
+    """
+    import database.users as users_db
+
+    user_profile = users_db.get_user_profile(uid)
+    user_name = user_profile.get('name') or 'User'
+
+    all_person_ids = set()
+    for conv in conversations:
+        for seg in conv.get('transcript_segments', []):
+            if seg.get('person_id'):
+                all_person_ids.add(seg['person_id'])
+
+    people_map = {}
+    if all_person_ids:
+        people_data = users_db.get_people_by_ids(uid, list(all_person_ids))
+        people_map = {p['id']: p['name'] for p in people_data}
+
+    for conv in conversations:
+        for seg in conv.get('transcript_segments', []):
+            if seg.get('is_user'):
+                seg['speaker_name'] = user_name
+            elif seg.get('person_id') and seg['person_id'] in people_map:
+                seg['speaker_name'] = people_map[seg['person_id']]
+            else:
+                seg['speaker_name'] = f"Speaker {seg.get('speaker_id', 0)}"
+
+
+def populate_folder_names(uid: str, conversations: List[Dict]) -> None:
+    """Add folder_name to conversations based on folder_id mappings.
+
+    Mutates conversation dicts in-place. Batch-loads all folder IDs in one query.
+    """
+    import database.folders as folders_db
+
+    folder_ids = set()
+    for conv in conversations:
+        if conv.get('folder_id'):
+            folder_ids.add(conv['folder_id'])
+
+    if not folder_ids:
+        for conv in conversations:
+            conv['folder_name'] = None
+        return
+
+    all_folders = folders_db.get_folders(uid)
+    folder_map = {f['id']: f['name'] for f in all_folders}
+
+    for conv in conversations:
+        folder_id = conv.get('folder_id')
+        conv['folder_name'] = folder_map.get(folder_id) if folder_id else None
+
+
+# ---------------------------------------------------------------------------
+# Redact: locked-content stripping
+# ---------------------------------------------------------------------------
+
+
+def redact_conversation_for_list(conv: Dict) -> Dict:
+    """Standard list-view redaction: strip detail fields, keep title/overview."""
+    if not conv.get('is_locked', False):
+        return conv
+    if 'structured' in conv:
+        conv['structured'] = (
+            dict(conv['structured']) if not isinstance(conv['structured'], dict) else conv['structured']
+        )
+        conv['structured']['action_items'] = []
+        conv['structured']['events'] = []
+    conv['apps_results'] = []
+    conv['plugins_results'] = []
+    conv['suggested_summarization_apps'] = []
+    conv['transcript_segments'] = []
+    return conv
+
+
+def redact_conversation_for_integration(conv: Dict) -> Dict:
+    """Integration-view redaction: strip everything including title/overview."""
+    if not conv.get('is_locked', False):
+        return conv
+    if 'structured' in conv:
+        conv['structured'] = (
+            dict(conv['structured']) if not isinstance(conv['structured'], dict) else conv['structured']
+        )
+        conv['structured']['title'] = ''
+        conv['structured']['overview'] = ''
+        conv['structured']['action_items'] = []
+        conv['structured']['events'] = []
+    conv['apps_results'] = []
+    conv['plugins_results'] = []
+    conv['suggested_summarization_apps'] = []
+    conv['transcript_segments'] = []
+    return conv
+
+
+def redact_conversations_for_list(conversations: List[Dict]) -> List[Dict]:
+    """Apply standard list redaction to a batch of conversations."""
+    return [redact_conversation_for_list(c) for c in conversations]
+
+
+def redact_conversations_for_integration(conversations: List[Dict]) -> List[Dict]:
+    """Apply integration redaction to a batch of conversations."""
+    return [redact_conversation_for_integration(c) for c in conversations]
+
+
+# ---------------------------------------------------------------------------
+# Serialize: datetime handling, dict conversion
+# ---------------------------------------------------------------------------
+
+
 def conversations_to_string(
     conversations: Sequence[Conversation],
     use_transcript: bool = False,
@@ -18,7 +136,7 @@ def conversations_to_string(
 ) -> str:
     """Format a sequence of Conversation objects into a human-readable string.
 
-    Callers must pass hydrated Conversation objects (use factory.hydrate_conversation
+    Callers must pass deserialized Conversation objects (use factory.deserialize_conversation
     for raw dicts). This function does NOT accept dicts.
     """
     result = []
@@ -83,12 +201,7 @@ def conversations_to_string(
 
 
 def serialize_datetimes(obj: Any) -> Any:
-    """Recursively convert datetime objects to ISO format strings.
-
-    Replaces: models/conversation.py::as_dict_cleaned_dates (nested helper),
-              utils/webhooks.py::_json_serialize_datetime,
-              utils/app_integrations.py::_json_serialize_datetime.
-    """
+    """Recursively convert datetime objects to ISO format strings."""
     if isinstance(obj, datetime):
         return obj.isoformat()
     elif isinstance(obj, dict):
@@ -99,9 +212,5 @@ def serialize_datetimes(obj: Any) -> Any:
 
 
 def conversation_to_dict(conversation: Conversation) -> Dict:
-    """Convert a Conversation to a JSON-safe dict with ISO datetime strings.
-
-    Replaces Conversation.as_dict_cleaned_dates(). Serialization is a rendering
-    concern, not a model concern.
-    """
+    """Convert a Conversation to a JSON-safe dict with ISO datetime strings."""
     return serialize_datetimes(conversation.dict())
