+++ backend/utils/retrieval/tools/conversation_tools.py
@@ -15,7 +15,7 @@ import database.users as users_db
 import database.vector_db as vector_db
 from models.conversation import Conversation
 from models.other import Person
-from utils.conversations.factory import hydrate_conversation
+from utils.conversations.factory import deserialize_conversation
 from utils.conversations.render import conversations_to_string
 from utils.llm.clients import embeddings
 import logging
@@ -223,7 +223,7 @@ def get_conversations_tool(
         conversations = []
         for conv_data in conversations_data:
             try:
-                conversation = hydrate_conversation(conv_data)
+                conversation = deserialize_conversation(conv_data)
 
                 # Limit transcript segments if needed (mimicking integration.py pattern)
                 if (
@@ -444,7 +444,7 @@ def search_conversations_tool(
         conversations = []
         for conv_data in conversations_data:
             try:
-                conversation = hydrate_conversation(conv_data)
+                conversation = deserialize_conversation(conv_data)
 
                 # Limit transcript segments if needed
                 if (
